+++ buildSrc/src/main/kotlin/JavaVersion.kt
@@ -1,7 +1,0 @@
-val javaVersion: String
-    get() = System.getProperty("java.version")!!
-
-val javaVersionMajor: Int
-    get() = javaVersion
-        .substringBefore(".")
-        .toInt()