+++ kotlinx-coroutines-core/common/src/flow/operators/Lint.kt
@@ -4,6 +4,7 @@
 
 package kotlinx.coroutines.flow
 
+import kotlinx.coroutines.*
 import kotlin.coroutines.*
 
 /**
@@ -41,3 +42,31 @@ public fun <T> StateFlow<T>.conflate(): Flow<T> = noImpl()
     replaceWith = ReplaceWith("this")
 )
 public fun <T> StateFlow<T>.distinctUntilChanged(): Flow<T> = noImpl()
+
+//@Deprecated(
+//    message = "isActive is resolved into the extension of outer CoroutineScope which is likely to be an error." +
+//        "Use currentCoroutineContext().isActive or cancellable() operator instead " +
+//        "or specify the receiver of isActive explicitly. " +
+//        "Additionally, flow {} builder emissions are cancellable by default.",
+//    level = DeprecationLevel.WARNING, // ERROR in 1.4
+//    replaceWith = ReplaceWith("currentCoroutineContext().isActive")
+//)
+//public val FlowCollector<*>.isActive: Boolean
+//    get() = noImpl()
+//
+//@Deprecated(
+//    message = "cancel() is resolved into the extension of outer CoroutineScope which is likely to be an error." +
+//        "Use currentCoroutineContext().cancel() instead or specify the receiver of cancel() explicitly",
+//    level = DeprecationLevel.WARNING, // ERROR in 1.4
+//    replaceWith = ReplaceWith("currentCoroutineContext().cancel(cause)")
+//)
+//public fun FlowCollector<*>.cancel(cause: CancellationException? = null): Unit = noImpl()
+//
+//@Deprecated(
+//    message = "coroutineContext is resolved into the property of outer CoroutineScope which is likely to be an error." +
+//        "Use currentCoroutineContext() instead or specify the receiver of coroutineContext explicitly",
+//    level = DeprecationLevel.WARNING, // ERROR in 1.4
+//    replaceWith = ReplaceWith("currentCoroutineContext()")
+//)
+//public val FlowCollector<*>.coroutineContext: CoroutineContext
+//    get() = noImpl()